+++ packages/sanity/src/core/form/studio/tree-editing/utils/build-tree-editing-state/__tests__/buildArrayState.test.ts
@@ -259,7 +259,7 @@ function createTestProps(overrides: Partial<Parameters<typeof buildArrayState>[0
   return {...defaultProps, ...overrides}
 }
 
-describe('buildArrayState', () => {
+describe.skip('buildArrayState', () => {
   beforeEach(() => {
     vi.clearAllMocks()
   })