+++ packages/sanity/src/core/releases/components/documentHeader/VersionChip.tsx
@@ -1,12 +1,6 @@
-import {type ReleaseDocument, type ReleaseState} from '@sanity/client'
+import {type ReleaseDocument} from '@sanity/client'
 import {ComposeSparklesIcon, LockIcon} from '@sanity/icons'
-import {
-  type BadgeTone,
-  Button, // eslint-disable-line no-restricted-imports
-  useClickOutsideEvent,
-  useGlobalKeyDown,
-  useToast,
-} from '@sanity/ui'
+import {type BadgeTone, useClickOutsideEvent, useGlobalKeyDown, useToast} from '@sanity/ui'
 import {
   memo,
   type MouseEvent,
@@ -18,7 +12,6 @@ import {
   useState,
 } from 'react'
 import {useObservable} from 'react-rx'
-import {styled} from 'styled-components'
 
 import {Popover, Tooltip} from '../../../../ui-components'
 import {useCanvasCompanionDocsStore} from '../../../canvas/store/useCanvasCompanionDocsStore'
@@ -29,23 +22,12 @@ import {getDraftId, getPublishedId, getVersionId} from '../../../util/draftUtils
 import {isCardinalityOneRelease} from '../../../util/releaseUtils'
 import {useVersionOperations} from '../../hooks/useVersionOperations'
 import {getReleaseIdFromReleaseDocumentId} from '../../util/getReleaseIdFromReleaseDocumentId'
+import {Chip} from '../Chip'
 import {DiscardVersionDialog} from '../dialog/DiscardVersionDialog'
 import {ReleaseAvatarIcon} from '../ReleaseAvatar'
 import {VersionContextMenu} from './contextMenu/VersionContextMenu'
 import {CopyToDraftsDialog} from './dialog/CopyToDraftsDialog'
 import {CopyToNewReleaseDialog} from './dialog/CopyToNewReleaseDialog'
-
-const ChipButtonContainer = styled.span`
-  display: inline-flex;
-  --border-color: var(--card-border-color);
-`
-
-const ChipButton = styled(Button)`
-  flex: none;
-  transition: none;
-  cursor: pointer;
-  --card-border-color: var(--border-color);
-`
 
 type VersionChipDialogState = 'idle' | 'discard-version' | 'create-release' | 'copy-to-drafts'
 
@@ -65,26 +47,29 @@ const useVersionIsLinked = (documentId: string, fromRelease: string) => {
   return companionDocs?.data.some((companion) => companion?.studioDocumentId === versionId)
 }
 
+const CONTEXT_MENU_CLOSED = {open: false as const}
+
 /**
  * @internal
  */
 export const VersionChip = memo(function VersionChip(props: {
   disabled?: boolean
   selected: boolean
-  tooltipContent: ReactNode
+  tooltipContent?: ReactNode
   onClick: () => void
   text: string
+  // if the VersionChip itself is contained in a portal (e.g., as in the NonReleaseVersionSelect)
+  // there is no need to also make the context menu a portal (and it also breaks useClickOutside)
+  contextMenuPortal?: boolean
   tone: BadgeTone
   locked?: boolean
   onCopyToDraftsNavigate: () => void
   contextValues: {
     documentId: string
+    documentType: string
     releases: ReleaseDocument[]
     releasesLoading: boolean
-    documentType: string
-    menuReleaseId: string
-    fromRelease: string
-    releaseState?: ReleaseState
+    bundleId: string
     isVersion: boolean
     disabled?: boolean
     isGoingToUnpublish?: boolean
@@ -97,6 +82,7 @@ export const VersionChip = memo(function VersionChip(props: {
     tooltipContent,
     onClick,
     text,
+    contextMenuPortal = true,
     tone,
     locked = false,
     onCopyToDraftsNavigate,
@@ -105,8 +91,7 @@ export const VersionChip = memo(function VersionChip(props: {
       releases,
       releasesLoading,
       documentType,
-      menuReleaseId,
-      fromRelease,
+      bundleId,
       isVersion,
       disabled: contextMenuDisabled = false,
       isGoingToUnpublish = false,
@@ -114,11 +99,11 @@ export const VersionChip = memo(function VersionChip(props: {
     },
   } = props
   const releasesToolAvailable = useReleasesToolAvailable()
-  const isLinked = useVersionIsLinked(documentId, fromRelease)
+  const isLinked = useVersionIsLinked(documentId, bundleId)
 
-  const [contextMenuPoint, setContextMenuPoint] = useState<{x: number; y: number} | undefined>(
-    undefined,
-  )
+  const [contextMenu, setContextMenu] = useState<
+    {open: true; translate: {x: number; y: number}} | {open: false}
+  >({open: false})
   const popoverRef = useRef<HTMLDivElement | null>(null)
   const [dialogState, setDialogState] = useState<VersionChipDialogState>('idle')
 
@@ -128,28 +113,27 @@ export const VersionChip = memo(function VersionChip(props: {
     if (selected) chipRef.current?.scrollIntoView({inline: 'center'})
   }, [selected])
 
-  const docId = isVersion ? getVersionId(documentId, fromRelease) : documentId // operations recognises publish and draft as empty
+  const docId = isVersion ? getVersionId(documentId, bundleId) : documentId // operations recognises publish and draft as empty
 
   const {createVersion} = useVersionOperations()
   const toast = useToast()
   const {t} = useTranslation()
 
-  const close = useCallback(() => setContextMenuPoint(undefined), [])
+  const close = useCallback(() => setContextMenu(CONTEXT_MENU_CLOSED), [])
+  const [referenceElement, setReferenceElement] = useState<HTMLElement | null>(null)
 
   const handleContextMenu = useCallback((event: MouseEvent<HTMLButtonElement>) => {
     event.preventDefault()
-
-    setContextMenuPoint({x: event.clientX, y: event.clientY})
+    const elementRect = event.currentTarget?.getBoundingClientRect()
+    setContextMenu({
+      open: true,
+      // note: this offsets the context menu popover position
+      // and depends on placement=bottom-start
+      translate: {x: event.clientX - elementRect.left, y: elementRect.top - event.clientY},
+    })
   }, [])
 
-  useClickOutsideEvent(
-    () => {
-      if (contextMenuPoint?.x && contextMenuPoint?.y) {
-        close()
-      }
-    },
-    () => [popoverRef.current],
-  )
+  useClickOutsideEvent(close, () => [popoverRef.current])
 
   useGlobalKeyDown(
     useCallback(
@@ -192,27 +176,6 @@ export const VersionChip = memo(function VersionChip(props: {
     [close, createVersion, docId, t, toast],
   )
 
-  const referenceElement = useMemo(() => {
-    if (!contextMenuPoint) {
-      return null
-    }
-
-    return {
-      getBoundingClientRect() {
-        return {
-          x: contextMenuPoint.x,
-          y: contextMenuPoint.y,
-          left: contextMenuPoint.x,
-          top: contextMenuPoint.y,
-          right: contextMenuPoint.x,
-          bottom: contextMenuPoint.y,
-          width: 0,
-          height: 0,
-        }
-      },
-    } as HTMLElement
-  }, [contextMenuPoint])
-
   const contextMenuHandler = disabled || !releasesToolAvailable ? undefined : handleContextMenu
 
   const isScheduledDraft = release && isVersion && isCardinalityOneRelease(release)
@@ -227,26 +190,21 @@ export const VersionChip = memo(function VersionChip(props: {
     <>
       <Tooltip content={tooltipContent} fallbackPlacements={[]} portal placement="bottom">
         {/* This span is needed to make the tooltip work in disabled buttons */}
-        <ChipButtonContainer>
-          <ChipButton
+        <span ref={chipRef}>
+          <Chip
             data-testid={`document-header-${text.replaceAll(' ', '-')}-chip`}
-            ref={chipRef}
+            ref={setReferenceElement}
             disabled={disabled}
             mode={disabled ? 'ghost' : 'bleed'}
             onClick={onClick}
             selected={selected}
             tone={tone}
             onContextMenu={contextMenuHandler}
-            paddingY={2}
-            paddingLeft={2}
-            paddingRight={3}
-            space={2}
-            radius="full"
             icon={<ReleaseAvatarIcon tone={tone} />}
             iconRight={isLinked ? <ComposeSparklesIcon /> : locked && <LockIcon />}
             text={text}
           />
-        </ChipButtonContainer>
+        </span>
       </Tooltip>
 
       <Popover
@@ -256,7 +214,7 @@ export const VersionChip = memo(function VersionChip(props: {
             documentId={documentId}
             releases={releases}
             releasesLoading={releasesLoading}
-            fromRelease={fromRelease}
+            fromRelease={bundleId}
             isVersion={isVersion}
             onDiscard={openDiscardDialog}
             onCreateRelease={openCreateReleaseDialog}
@@ -273,22 +231,23 @@ export const VersionChip = memo(function VersionChip(props: {
           />
         }
         fallbackPlacements={[]}
-        open={Boolean(referenceElement)}
-        portal
+        open={contextMenu.open}
+        portal={contextMenuPortal}
         placement="bottom-start"
         ref={popoverRef}
         referenceElement={referenceElement}
         zOffset={10}
+        style={
+          contextMenu.open
+            ? {transform: `translate(${contextMenu.translate.x}px, ${contextMenu.translate.y}px)`}
+            : undefined
+        }
       />
 
       {dialogState === 'discard-version' && (
         <DiscardVersionDialog
           onClose={() => setDialogState('idle')}
-          documentId={
-            isVersion
-              ? getVersionId(documentId, getReleaseIdFromReleaseDocumentId(menuReleaseId))
-              : documentId
-          }
+          documentId={isVersion ? getVersionId(documentId, bundleId) : documentId}
           fromPerspective={text}
           documentType={documentType}
         />
@@ -298,11 +257,7 @@ export const VersionChip = memo(function VersionChip(props: {
         <CopyToNewReleaseDialog
           onClose={() => setDialogState('idle')}
           onCreateVersion={handleAddVersion}
-          documentId={
-            isVersion
-              ? getVersionId(documentId, getReleaseIdFromReleaseDocumentId(menuReleaseId))
-              : documentId
-          }
+          documentId={isVersion ? getVersionId(documentId, bundleId) : documentId}
           documentType={documentType}
           tone={tone}
           title={text}
@@ -313,7 +268,7 @@ export const VersionChip = memo(function VersionChip(props: {
         <CopyToDraftsDialog
           onClose={() => setDialogState('idle')}
           documentId={documentId}
-          fromRelease={fromRelease}
+          fromRelease={bundleId}
           onNavigate={onCopyToDraftsNavigate}
         />
       )}