+++ packages/sanity/src/core/releases/components/documentHeader/contextMenu/ScheduledDraftContextMenu.tsx
@@ -8,13 +8,14 @@ import {MenuItem} from '../../../../../ui-components'
 import {useTranslation} from '../../../../i18n/hooks/useTranslation'
 import {type UseScheduledDraftMenuActionsReturn} from '../../../../singleDocRelease/hooks/useScheduledDraftMenuActions'
 import {RELEASES_SCHEDULED_DRAFTS_INTENT} from '../../../../singleDocRelease/plugin'
+import {useWorkspace} from '../../../../studio'
 import {isReleaseScheduledOrScheduling} from '../../../util/util'
 import {useHasCopyToDraftOption} from './CopyToDraftsMenuItem'
 import {CopyToReleaseMenuGroup} from './CopyToReleaseMenuGroup'
 
 interface ScheduledDraftContextMenuProps {
   releases: ReleaseDocument[]
-  fromRelease: string
+  bundleId: string
   onCreateRelease: () => void
   onCopyToDrafts: () => void
   onCopyToDraftsNavigate: () => void
@@ -25,6 +26,7 @@ interface ScheduledDraftContextMenuProps {
   scheduledDraftMenuActions: UseScheduledDraftMenuActionsReturn
   documentId: string
   documentType: string
+  release?: ReleaseDocument
 }
 
 export const ScheduledDraftContextMenu = memo(function ScheduledDraftContextMenu(
@@ -32,7 +34,7 @@ export const ScheduledDraftContextMenu = memo(function ScheduledDraftContextMenu
 ) {
   const {
     releases,
-    fromRelease,
+    bundleId,
     onCreateRelease,
     onCopyToDrafts,
     onCopyToDraftsNavigate,
@@ -45,11 +47,12 @@ export const ScheduledDraftContextMenu = memo(function ScheduledDraftContextMenu
     documentType,
   } = props
   const {t} = useTranslation()
-  const hasCopyToDraftOption = useHasCopyToDraftOption(documentType, fromRelease)
+  const hasCopyToDraftOption = useHasCopyToDraftOption(documentType, bundleId)
 
   const isCopyToReleaseDisabled = disabled || !hasCreatePermission || isGoingToUnpublish
   const copyToReleaseOptions = releases.filter((r) => !isReleaseScheduledOrScheduling(r))
-  const showCopyToReleaseMenuItem = copyToReleaseOptions.length > 0 || hasCopyToDraftOption
+  const isReleasesEnabled = !!useWorkspace().releases?.enabled
+  const showCopyToReleaseMenuItem = isReleasesEnabled && copyToReleaseOptions.length > 0
 
   const {actions} = scheduledDraftMenuActions
 
@@ -66,11 +69,13 @@ export const ScheduledDraftContextMenu = memo(function ScheduledDraftContextMenu
         <MenuItem icon={CalendarIcon} text={t('release.action.view-scheduled-drafts')} />
       </IntentLink>
       <MenuDivider />
-      {showCopyToReleaseMenuItem && (
+      {(showCopyToReleaseMenuItem || hasCopyToDraftOption) && (
         <>
           <CopyToReleaseMenuGroup
             releases={copyToReleaseOptions}
-            fromRelease={fromRelease}
+            bundleId={bundleId}
+            hasCopyToDraftOption={hasCopyToDraftOption}
+            isReleasesEnabled={isReleasesEnabled}
             onCreateRelease={onCreateRelease}
             onCopyToDrafts={onCopyToDrafts}
             onCopyToDraftsNavigate={onCopyToDraftsNavigate}