+++ packages/sanity/src/structure/panes/document/DocumentPaneProvider.tsx
@@ -289,7 +289,7 @@ export function DocumentPaneProvider(props: DocumentPaneProviderProps) {
 
   const {data: releases = []} = useActiveReleases()
 
-  const actionsPerspective = useMemo(
+  const actionsVersionType = useMemo(
     () =>
       getDocumentVersionType(
         params,
@@ -302,20 +302,20 @@ export function DocumentPaneProvider(props: DocumentPaneProviderProps) {
     [params, selectedReleaseId, value, selectedPerspectiveName, draftsEnabled, releases],
   )
 
-  const documentActionsProps: PartialContext<DocumentActionsContext> = useMemo(
+  const documentActionsContext: PartialContext<DocumentActionsContext> = useMemo(
     () => ({
       schemaType: documentType,
       documentId,
-      versionType: actionsPerspective,
+      versionType: actionsVersionType,
       releaseId: selectedReleaseId,
     }),
-    [documentType, documentId, actionsPerspective, selectedReleaseId],
+    [documentType, documentId, actionsVersionType, selectedReleaseId],
   )
 
   // Resolve document actions
   const actions = useMemo(
-    () => documentActions(documentActionsProps),
-    [documentActions, documentActionsProps],
+    () => documentActions(documentActionsContext),
+    [documentActions, documentActionsContext],
   )
 
   const handlePathOpen = useCallback(